requirements.txt
numpy
finta
python-binance
matplotlib